cogs implementation
